Fix TS exports of enums from BBUI, bump portal version
